+++ levenshtein/csharp/in-process/code.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net9.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <Compile Include="..\..\..\lib\csharp\benchmark.cs">
+            <Link>benchmark.cs</Link>
+        </Compile>
+    </ItemGroup>
+
+</Project>
