+++ levenshtein/csharp/legacy/code.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
@@ -6,11 +6,5 @@
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Include="..\..\lib\csharp\Benchmark.cs">
-      <Link>Benchmark.cs</Link>
-    </Compile>
-  </ItemGroup>
 
 </Project>
